Attempt some testing for login and sign up interfaces
Some attempts at doing the login and sign up interface but I keep getting the below error:

Whitelabel Error Page

This application has no explicit mapping for /error, so you are seeing this as a fallback.
<date and time>
There was an unexpected error (type=Not Found, status=404).
No static resource <name of html it is supposed to reference>.