Added the session, login and logout functionality
Currently login will accept any username that is not blank and any password with 8+ characters as valid since the database as of now is not implemented yet.
The system will also redirect you away from the 'login' and 'signup' pages back to the 'main' page if you are already logged in.